Add AP2 datacenter
Add the AP2 datacenter to the cdk constructs site list.  Move the siteList
constant to the top level since it is duplicated between the lambda and
fargate files, as well as the invalid site error.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-6912
