+++ src/datadog-lambda.ts
@@ -27,6 +27,8 @@ import {
   Transport,
   applyExtensionLayer,
   DD_TAGS,
+  siteList,
+  invalidSiteError,
 } from "./index";
 import { LambdaFunction } from "./interfaces";
 import { setTags } from "./tag";
@@ -272,23 +274,13 @@ export function validateProps(props: DatadogLambdaProps, apiKeyArnOverride = fal
   log.debug("Validating props...");
 
   checkForMultipleApiKeys(props, apiKeyArnOverride);
-  const siteList: string[] = [
-    "datadoghq.com",
-    "datadoghq.eu",
-    "us3.datadoghq.com",
-    "us5.datadoghq.com",
-    "ap1.datadoghq.com",
-    "ddog-gov.com",
-  ];
   if (
     props.site !== undefined &&
     !siteList.includes(props.site.toLowerCase()) &&
     !(props.site.startsWith("${Token[") && props.site.endsWith("]}")) &&
     !process.env.DD_CDK_BYPASS_SITE_VALIDATION
   ) {
-    throw new Error(
-      "Warning: Invalid site URL. Must be either datadoghq.com, datadoghq.eu, us3.datadoghq.com, us5.datadoghq.com, ap1.datadoghq.com, or ddog-gov.com.",
-    );
+    throw new Error(invalidSiteError);
   }
 
   if (
